Forms refactoring - Main People Form
With permission from Emily, Emi and Martin, I was able to delete all the complication in the main people form connected with the business requirement of showing a `At least two partners are needed` error message.
I have also killed inheritance and composition classes used there (for now) as I found them unecessary.
This also removes the people type.
