+++ app/forms/waste_exemptions_engine/main_people_form.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 module WasteExemptionsEngine
-  class MainPeopleForm < PersonForm
-    include CanLimitNumberOfMainPeople
-
+  class MainPeopleForm < BaseForm
     attr_accessor :business_type
+    attr_accessor :first_name, :last_name
 
-    validates_with MainPersonValidator
+    delegate :transient_people, to: :transient_registration
+
+    validates :first_name, :last_name, "waste_exemptions_engine/person_name": true
 
     def initialize(transient_registration)
       super
@@ -14,11 +15,17 @@ module WasteExemptionsEngine
       self.business_type = @transient_registration.business_type
 
       # If there's only one main person, we can pre-fill the fields so users can easily edit them
-      prefill_form if can_only_have_one_main_person? && @transient_registration.transient_people.present?
+      prefill_form if @transient_registration.transient_people.count == 1
     end
 
-    def person_type
-      TransientPerson.person_types[:partner]
+    def submit(params)
+      # Assign the params for validation and pass them to the BaseForm method for updating
+      self.first_name = params[:first_name]
+      self.last_name = params[:last_name]
+
+      set_up_new_person if valid?
+
+      super({})
     end
 
     private
@@ -28,5 +35,11 @@ module WasteExemptionsEngine
       self.last_name = @transient_registration.transient_people.first.last_name
     end
 
+    def set_up_new_person
+      @transient_registration.transient_people.build(
+        first_name: first_name,
+        last_name: last_name
+      )
+    end
   end
 end
