+++ app/models/waste_exemptions_engine/person.rb
@@ -6,8 +6,5 @@ module WasteExemptionsEngine
     self.table_name = "people"
 
     belongs_to :registration
-
-    enum person_type: { partner: 0 }
-
   end
 end
