+++ app/views/waste_exemptions_engine/main_people_forms/new.html.erb
@@ -1,62 +1,49 @@
 <%= render("waste_exemptions_engine/shared/back", back_path: back_main_people_forms_path(@main_people_form.token)) %>
 
-<% if @main_people_form.can_only_have_one_main_person? %>
-<div class="text">
-<% else %>
 <div class="grid-row">
   <div class="column-two-thirds">
-<% end %>
+    <%= form_for(@main_people_form) do |f| %>
+      <%= render("waste_exemptions_engine/shared/errors", object: @main_people_form) %>
 
-<%= form_for(@main_people_form) do |f| %>
-  <%= render("waste_exemptions_engine/shared/errors", object: @main_people_form) %>
+      <h1 class="heading-large"><%= t(".heading.#{@main_people_form.business_type}") %></h1>
 
-  <h1 class="heading-large"><%= t(".heading.#{@main_people_form.business_type}") %></h1>
+      <p><%= t(".description_1.#{@main_people_form.business_type}") %></p>
 
-  <p><%= t(".description_1.#{@main_people_form.business_type}") %></p>
+      <%= render("waste_exemptions_engine/shared/person_name", form: @main_people_form, f: f) %>
 
-  <%= render("waste_exemptions_engine/shared/person_name", form: @main_people_form, f: f) %>
+      <%= f.hidden_field :token, value: @main_people_form.token %>
 
-  <%= f.hidden_field :token, value: @main_people_form.token %>
-
-  <% if @main_people_form.can_only_have_one_main_person? %>
-    <div class="form-group">
-      <%= f.submit t(".next_button"), class: "button" %>
-    </div>
-  <% elsif @main_people_form.minimum_main_people > (@main_people_form.number_of_existing_main_people + 1) %>
-    <div class="form-group">
-      <%= f.submit t(".add_person_link"), class: "button" %>
-    </div>
-  <% else %>
-    <div class="form-group">
-      <%= f.submit t(".add_person_link"), class: "button-link" %>
-    </div>
-    <div class="form-group">
-      <%= f.submit t(".next_button"), class: "button" %>
-    </div>
-  <% end %>
-<% end %>
-
-<% if @main_people_form.can_only_have_one_main_person? %>
-  </div> <!-- closes .text -->
-<% else %>
-  </div> <!-- closes .column-two-thirds -->
-  <% if @main_people_form.number_of_existing_main_people > 0 %>
-  <div class="column-one-third">
-    <h2 class="heading-small"><%= t(".list_of_people") %></h2>
-
-    <ul class="main-person-list">
-    <% @transient_registration.transient_people.each do |person| %>
-        <li>
-          <%= person.first_name %> <%= person.last_name %>
-          <%= button_to(t(".delete_person_link"),
-              delete_person_main_people_forms_path(person.id),
-              class: "button-link",
-              method: :delete,
-              params: { token: @main_people_form.token }) %>
-        </li>
+      <% if @main_people_form.transient_people.count < 1 %>
+        <div class="form-group">
+          <%= f.submit t(".add_person_link"), class: "button" %>
+        </div>
+      <% else %>
+        <div class="form-group">
+          <%= f.submit t(".add_person_link"), class: "button-link" %>
+        </div>
+        <div class="form-group">
+          <%= f.submit t(".next_button"), class: "button" %>
+        </div>
+      <% end %>
     <% end %>
-    </ul>
-  </div> <!-- closes .column-one-third -->
+  </div> <!-- closes .column-two-thirds -->
+
+  <% if @main_people_form.transient_people.count > 0 %>
+    <div class="column-one-third">
+      <h2 class="heading-small"><%= t(".list_of_people") %></h2>
+
+      <ul class="main-person-list">
+        <% @transient_registration.transient_people.each do |person| %>
+          <li>
+            <%= person.first_name %> <%= person.last_name %>
+            <%= button_to(t(".delete_person_link"),
+                delete_person_main_people_forms_path(person.id),
+                class: "button-link",
+                method: :delete,
+                params: { token: @main_people_form.token }) %>
+          </li>
+        <% end %>
+      </ul>
+    </div> <!-- closes .column-one-third -->
   <% end %>
 </div> <!-- closes .grid-row -->
-<% end %>
