+++ spec/forms/waste_exemptions_engine/main_people_form_spec.rb
@@ -6,17 +6,6 @@ module WasteExemptionsEngine
   RSpec.describe MainPeopleForm, type: :model do
     subject(:form) { build(:main_people_form) }
 
-    it "should be a type of PersonForm" do
-      expect(described_class.ancestors)
-        .to include(WasteExemptionsEngine::PersonForm)
-    end
-
-    it "validates the person using the MainPersonValidator class" do
-      validators = form._validators
-      validator_classes = validators.values.flatten.map(&:class)
-      expect(validator_classes).to include(WasteExemptionsEngine::MainPersonValidator)
-    end
-
     it_behaves_like "a validated form", :main_people_form do
       let(:valid_params) do
         { token: form.token, first_name: "Joe", last_name: "Bloggs" }
@@ -25,7 +14,8 @@ module WasteExemptionsEngine
         [
           { token: form.token, first_name: "", last_name: "Bloggs" },
           { token: form.token, first_name: "Joe", last_name: "" },
-          { token: form.token, first_name: "", last_name: "" }
+          { token: form.token, first_name: "", last_name: "" },
+          { token: form.token, first_name: "Bob", last_name: "This Name Is Far Too Long 1 .... 36!" }
         ]
       end
     end
@@ -45,7 +35,6 @@ module WasteExemptionsEngine
           person = transient_registration.transient_people.first
           expect(person.first_name).to eq(first_name)
           expect(person.last_name).to eq(last_name)
-          expect(person.person_type).to eq(person_type)
         end
       end
     end
